+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 1.1.0
+
+  * Compatibility with servant-0.12.
+  * Added instance for WebSocketPending.
+
 # Version 1.0.0
 
   * Initial release of `servant-websockets`.